Adjust member count data type

+++ pkg/monitor/members/members.go
@@ -11,7 +11,7 @@ import (
 
 type membersCache struct {
 	mu         sync.RWMutex
-	numMembers float64
+	numMembers int
 }
 
 var (
@@ -46,10 +46,10 @@ func check(dgBotSession *discordgo.Session) {
 	defer cache.mu.Unlock()
 
 	checkNum := cache.numMembers
-	numMembers := float64(0)
+	numMembers := 0
 
 	for _, guild := range dgBotSession.State.Guilds {
-		numMembers += float64(guild.MemberCount)
+		numMembers += guild.MemberCount
 	}
 
 	if numMembers == checkNum {
@@ -59,7 +59,7 @@ func check(dgBotSession *discordgo.Session) {
 	update(numMembers)
 }
 
-func update(numMembers float64) {
+func update(numMembers int) {
 	cache.numMembers = numMembers
-	prometheusMembersGauge.Set(cache.numMembers)
+	prometheusMembersGauge.Set(float64(cache.numMembers))
 }
